Add inference results
These are all (raw) results from inference time with the participants. There is one file included which carries my name - it was an example of how different my scores are to those of other new participants to the game - proving the bias I would have brought simply by remembering the order of target grids.