Add files via upload
All scripts I used for fine-tuning the LLMs.
Merge adapters merges adapter tensors into the model.
process results processes the raw data of inferences and results in a single json file with all episodes included (includes u. a. role, ips, etc.)
requirements.txt includes the list of libraries that were installed in the vrtual env
train_eval-samples includes samples taken during the fine-tuning to gain insight into what the model is generating for the early stopping comprehension acc
woking-finetuning is the finee-tuning script to train the LLM. _llms is a slight adaptation for the data which has only model-model data.